parser/requirements.txt
redis
psycopg2-binary
pytesseract
Pillow
python-dotenv